NPE in WallpaperHandler
In WallpaperHandler, BitmapFactory.decodeStream() is used to create
a Bitmap that is then set by WallpaperManager as a wallpaper.
This method has null as return value when something fails as
decoding, resulting in a nullpointer exception in WallpaperManager.
Adding nullcheck on the Bitmap before setting it as wallpaper to
avoid the NPE.

Change-Id: Ib1a0f9331898162b21094c3ebe80bc21d66f8bd7