Pause WebView timer when Browser is paused
If the Browser is paused while loading a page it
shall finish loading the page before the WebView
timer is paused.

The WebView timer shall be paused in onProgressChanged
when the progress has reached 100 not from onPageFinished
since onPageFinished is called when the main frame is loaded
and at that time subframe and other resources might haven't
been finished loaded i.e. the page is still in load, which the
result in WebView timers are not paused.

With this patch the Browser won't consume system resources
if it's minimized while doing a page load.

Change-Id: Id971dbcedfbedb5b0e56cba4076cd724d590ac85